Remove db package and move classes to more appropriate locations
