+++ ethereum/core/src/main/java/tech/pegasys/pantheon/ethereum/storage/StorageProvider.java
@@ -12,7 +12,7 @@
  */
 package tech.pegasys.pantheon.ethereum.storage;
 
-import tech.pegasys.pantheon.ethereum.db.BlockchainStorage;
+import tech.pegasys.pantheon.ethereum.chain.BlockchainStorage;
 import tech.pegasys.pantheon.ethereum.mainnet.ProtocolSchedule;
 import tech.pegasys.pantheon.ethereum.worldstate.WorldStateStorage;
