+++ ethereum/core/src/test/java/tech/pegasys/pantheon/ethereum/chain/GenesisBlockMismatchTest.java
@@ -10,7 +10,7 @@
  * an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
  * specific language governing permissions and limitations under the License.
  */
-package tech.pegasys.pantheon.ethereum.db;
+package tech.pegasys.pantheon.ethereum.chain;
 
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatExceptionOfType;
